+++ scratch.js
@@ -7,23 +7,16 @@ const wtf = require('./src/index');
 // const doc = await wtf.fetch('Berlin', 'de');
 // console.log(doc.sentences(0).plaintext());
 // })();
-
-let str = ` {| class="oopsie"
-| first row
+let str = `{|class="wikitable sortable"
+!Name and Surname!!Height
 |-
-| Secod row
-{|
+|data-sort-value="Smith, John"|John Smith||1.85
 |-
-| embed 1
+|data-sort-value="Ray, Ian"|Ian Ray||1.89
 |-
-| embed 2
-|}
+|data-sort-value="Bianchi, Zachary"|Zachary Bianchi||1.72
 |-
-| Berlin!
-|-
-|}
-
-Actual first sentence  is here`;
-let text = wtf(str).sentences(0).text();
-console.log(text);
-// console.log(wtf(str).images());
+!Average:||1.82
+|}`;
+let doc = wtf(str);
+console.log(doc.tables(0).data[0]);
